requirements.txt
# TeleCursor Python Dependencies
# Install with: pip install -r requirements.txt

# Core
torch>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
pytest>=7.0.0

# Privacy & Differential Privacy
# For DP-SGD training (optional)
# opacus>=1.4.0  # Uncomment if using PyTorch DP library

# Model utilities
# flash-attn>=2.0.0  # Uncomment for FlashAttention support (requires CUDA)

# Data processing
pandas>=2.0.0

# Logging and monitoring
tensorboard>=2.14.0